Feedback improvements
